[refactor] 객체 세팅값 세팅 폴더에 보관
