+++ src/main/java/racingcar/domain/Car.java
@@ -2,7 +2,7 @@ package racingcar.domain;
 
 
 import java.util.Objects;
-import static racingcar.domain.CarSetting.*;
+import static racingcar.setting.CarSetting.*;
 import static racingcar.exception.CarExceptionMessage.*;
