+++ src/main/java/racingcar/setting/RaceRoundSetting.java
@@ -1,10 +1,9 @@
-package racingcar.domain;
+package racingcar.setting;
 
 public enum RaceRoundSetting {
 
     MIN_TOTAL_GAME_ROUND_COUNT(1),
     INIT_CURRENT_GAME_ROUND_COUNT(0);
-
 
     private final int setting;
